requirements.txt
# Python requirements for AFI RAG Pipeline
pdfplumber>=0.7.6
pandas>=1.5.0
openai>=1.0.0
chromadb>=0.4.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
jinja2>=3.1.0
tiktoken>=0.5.0